Shuffle custom placeholder directories
We had a part of the code that made placeholder directories. Let's
remove most of them and move the rest around.

The `app/views/pages` directory is used for static pages as used by
`high_voltage`. Move that into the `StaticGenerator`.

The `spec/support` directory is used for all tests, and implied by
creating subdirectories. The `TestingGenerator` does this for us.

This removes `spec/lib`, `spec/controllers`, `spec/helpers`,
`spec/support/matchers`, `spec/support/mixins`, and
`spec/support/shared_examples`.

The `spec/lib` and `spec/helpers` directories are for tests under `lib`
and `app/helpers`, respectively. These directories are rarely used these
days, so let's not make these directories by default.

The `spec/controllers` directory is for controller unit tests. These
have mostly been replaced by `spec/features`, `spec/requests`, and
`spec/system`. Let's hold off on this directory for now.

The subdirectories under `spec/support` -- `matchers`, `mixins`, and
`shared_examples` -- are far too often overlooked, and instead
everything is dropped directly under `spec/support`. And in practice,
90% of the time, this is fine. People just don't extract a ton of stuff
into `spec/support`. If a project does, they can find a way to
re-organize things as appropriate.

---

During the creation of this diff, a `suspenders:placeholders` generator
was extracted. Eventually it only created `app/services` and
`spec/services`, which lead to a discussion about whether we want to
encourage such behavior. The decision was made not to make those
directories in here (since we hadn't been doing that already) and
instead to move the discussion to our guides.
