+++ lib/suspenders/generators/base.rb
@@ -15,6 +15,15 @@ module Suspenders
       def app_name
         Rails.app_class.module_parent_name.demodulize.underscore.dasherize
       end
+
+      def empty_directory_with_keep_file(destination)
+        empty_directory(destination, {})
+        keep_file(destination)
+      end
+
+      def keep_file(destination)
+        create_file(File.join(destination, ".keep"))
+      end
     end
   end
 end
