+++ lib/suspenders/generators/static_generator.rb
@@ -6,5 +6,9 @@ module Suspenders
       gem "high_voltage"
       Bundler.with_clean_env { run "bundle install" }
     end
+
+    def make_placeholder_directory
+      empty_directory_with_keep_file "app/views/pages"
+    end
   end
 end
